Make buildDocsHelper a module for sharing eventually with elix.org
